+++ src/timeline/EventStore.h
@@ -29,7 +29,7 @@ public:
                 friend uint qHash(const Index &i, uint seed = 0) noexcept
                 {
                         QtPrivate::QHashCombine hash;
-                        seed = hash(seed, QByteArray::fromRawData(i.room.data(), i.room.size()));
+                        seed = hash(seed, QByteArray::fromRawData(i.room.data(), (int)i.room.size()));
                         seed = hash(seed, i.idx);
                         return seed;
                 }
@@ -46,8 +46,8 @@ public:
                 friend uint qHash(const IdIndex &i, uint seed = 0) noexcept
                 {
                         QtPrivate::QHashCombine hash;
-                        seed = hash(seed, QByteArray::fromRawData(i.room.data(), i.room.size()));
-                        seed = hash(seed, QByteArray::fromRawData(i.id.data(), i.id.size()));
+                        seed = hash(seed, QByteArray::fromRawData(i.room.data(), (int)i.room.size()));
+                        seed = hash(seed, QByteArray::fromRawData(i.id.data(), (int)i.id.size()));
                         return seed;
                 }
